Set Target Hours to Zero when date is holiday issue #47
